requirements.txt
pydantic
argparse
PyYAML
dependency-injector
Pint
